pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "isage-rag"
version = "0.3.0"
description = "SAGE RAG Framework - Document loaders, chunkers, retrievers for RAG pipelines"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "IntelliStream Team", email = "shuhao_zhang@hust.edu.cn" }]
keywords = [
    "RAG",
    "retrieval",
    "document-loader",
    "chunker",
    "retriever",
    "LLM",
    "AI",
]

dependencies = [
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    # SAGE layer dependencies (required for operators)
    "isage-common>=0.2.4.30",
    "isage-kernel>=0.2.4.28",
    "isage-libs>=0.2.4.22",
    # Runtime deps for operators
    "numpy>=1.21.0",
    "requests>=2.28.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
# Vector store backends
chroma = ["chromadb>=0.4.0"]
milvus = ["pymilvus>=2.3.0"]
backends = ["isage-rag[chroma,milvus]"]
# Operator optional deps
operators = [
    "openai>=1.0.0",
    "rouge>=1.0.0",
    "scikit-learn>=1.0.0",
    "feedparser>=6.0.0",
]
# middleware VDB backends (optional — lazy-loaded at runtime)
middleware = ["isage-middleware>=0.2.4.0"]
full = ["isage-rag[backends,operators,middleware]"]
dev = ["isage-rag[full]", "pytest>=7.0.0", "pytest-asyncio>=0.21.0", "ruff>=0.4.0"]

[project.urls]
"Homepage" = "https://github.com/intellistream/sage-rag"
"Repository" = "https://github.com/intellistream/sage-rag"

[tool.setuptools.packages.find]
where = ["src"]
include = ["sage_rag*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
